chore: move pkg/condition to internal
