+++ internal/condition/condition.go
@@ -9,8 +9,8 @@ import (
 	"github.com/google/cel-go/common"
 	celtypes "github.com/google/cel-go/common/types"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
+	"github.com/openfga/openfga/internal/condition/types"
 	"github.com/openfga/openfga/internal/errors"
-	"github.com/openfga/openfga/pkg/condition/types"
 	"golang.org/x/exp/maps"
 )
