+++ internal/condition/condition_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
-	"github.com/openfga/openfga/pkg/condition"
-	"github.com/openfga/openfga/pkg/condition/types"
+	"github.com/openfga/openfga/internal/condition"
+	"github.com/openfga/openfga/internal/condition/types"
 	"github.com/stretchr/testify/require"
 )
