+++ internal/condition/errors.go
