+++ internal/condition/types/converters.go
