+++ internal/condition/types/encoding.go
