+++ internal/condition/types/generic.go
