+++ internal/condition/types/primitives.go
