+++ internal/condition/types/types.go
