+++ internal/condition/types/types_test.go
