+++ pkg/typesystem/typesystem.go
@@ -10,7 +10,7 @@ import (
 	"sort"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
-	"github.com/openfga/openfga/pkg/condition"
+	"github.com/openfga/openfga/internal/condition"
 	"github.com/openfga/openfga/pkg/tuple"
 	"go.opentelemetry.io/otel"
 )
